Refactor: Complete project refactoring and simplification
This commit performs a full refactoring of the asset based on user request. The goals were to reduce file count, simplify the project structure, and add a sample placeholder.

Changes include:
- **Script Consolidation:** All runtime logic has been merged into a single `CensorEffect.cs` script. The redundant `CensorEffectBuiltin.cs` file and its corresponding nested class have been removed.
- **Directory Simplification:** The `Runtime/Builtin` and `Runtime/Common` subdirectories have been removed. All scripts and shaders now reside directly in the `Runtime` folder.
- **Unified Assembly:** The two previous assembly definition files have been replaced with a single, unified `.asmdef` file for the entire runtime module.
- **Added Samples Placeholder:** A `Samples~` directory has been added with a `README.md` file containing instructions for the user to create their own sample scene.
