+++ README.md
@@ -1,13 +1,12 @@
 ﻿# Censor Effect for Unity
 
-A versatile and performant camera effect to censor objects on a specific layer with a pixelated shader. Supports both the **Built-in Render Pipeline** and the **Universal Render Pipeline (URP)**.
+A versatile and performant camera effect to censor objects on a specific layer with a pixelated shader. Supports the **Built-in Render Pipeline**.
 
 ![Sample](https://github.com/user-attachments/assets/479ff24e-876b-4243-9fb5-2cf481f04a9c)
 
 ## Features
 - **Easy to use:** Add a component to your camera and you're ready to go.
 - **Performant:** Uses a two-pass blur for efficient area expansion.
-- **URP & Built-in RP:** Automatically adapts to your project's render pipeline.
 - **Occlusion Culling:** Optional depth testing to correctly hide censored objects behind others.
 
 ## How to Install
@@ -22,7 +21,6 @@ A versatile and performant camera effect to censor objects on a specific layer w
 
 ## How to Use
 
-### 1. Basic Setup (for all pipelines)
 1.  **Create a Censor Layer:**
     *   Go to `Edit` > `Project Settings` > `Tags and Layers`.
     *   Under `Layers`, add a new layer (e.g., "Censored").
@@ -35,17 +33,9 @@ A versatile and performant camera effect to censor objects on a specific layer w
     *   Select the `Camera` GameObject in your scene.
     *   In the Inspector, click `Add Component` and search for `CensorEffect`. Add it to the camera.
 
-### 2. Pipeline-Specific Setup
+The effect is now active! No other setup is required for the Built-in Render Pipeline.
 
-#### For Built-in Render Pipeline
-No extra steps are needed! The effect will automatically detect the pipeline and configure itself.
-
-#### For Universal Render Pipeline (URP)
-1.  Select your **URP Renderer** asset (usually found in your project's `Settings` folder, assigned to your URP Asset).
-2.  In the Inspector for the Renderer, click the **`Add Renderer Feature`** button.
-3.  Select the **`Censor Feature`** from the list.
-
-### 3. Configure the Censor Effect
+## Configure the Censor Effect
 -   Select your `Camera` GameObject.
 -   In the `CensorEffect` component, you can now configure the following settings:
     -   **Censor Layer:** The layer containing the objects to be pixelated.
