--- Runtime/Common/Resources/CensorEffect.shader.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: f96a40a257073de4498e95601672921f
-ShaderImporter:
-  externalObjects: {}
-  defaultTextures: []
-  nonModifiableTextures: []
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
