--- Runtime/Common/ineedmypills.CensorEffect.Runtime.Common.asmdef
@@ -1,15 +1,0 @@
-﻿
-{
-    "name": "ineedmypills.CensorEffect.Runtime.Common",
-    "rootNamespace": "CensorEffect.Runtime",
-    "references": [],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": false,
-    "defineConstraints": [],
-    "versionDefines": [],
-    "noEngineReferences": false
-}
