--- Runtime/URP/ineedmypills.CensorEffect.Runtime.URP.asmdef
@@ -1,25 +1,0 @@
-{
-    "name": "ineedmypills.CensorEffect.Runtime.URP",
-    "rootNamespace": "CensorEffect.Runtime.URP",
-    "references": [
-        "ineedmypills.CensorEffect.Runtime.Common",
-        "Unity.RenderPipelines.Universal.Runtime"
-    ],
-    "includePlatforms": [],
-    "excludePlatforms": [],
-    "allowUnsafeCode": false,
-    "overrideReferences": false,
-    "precompiledReferences": [],
-    "autoReferenced": true,
-    "defineConstraints": [
-        "UNITY_PIPELINE_URP"
-    ],
-    "versionDefines": [
-        {
-            "name": "com.unity.render-pipelines.universal",
-            "expression": "7.0.0",
-            "define": "UNITY_PIPELINE_URP"
-        }
-    ],
-    "noEngineReferences": false
-}
