+++ package.json
@@ -2,7 +2,7 @@
   "name": "com.ineedmypills.censor-effect",
   "version": "2.0.0",
   "displayName": "Censor Effect",
-  "description": "A versatile and performant camera effect to censor objects on a specific layer with a pixelated shader. Supports Built-in and Universal Render Pipeline (URP), with depth-based occlusion and a smooth expansion effect.",
+  "description": "A versatile and performant camera effect to censor objects on a specific layer with a pixelated shader. Supports the Built-in Render Pipeline, with depth-based occlusion and a smooth expansion effect.",
   "unity": "2019.4",
   "keywords": [
     "censor",
